requirements.txt
django~=6.0
pillow~=12.0.0
django-crispy-forms~=2.5
crispy-bulma~=0.12.0
django-environ~=0.12.0
django-email-validation~=2.1.1
django-simple-captcha~=0.6.3
daphne~=4.2.1
supervisor~=4.3.0
channels~=4.3.2
channels_redis~=4.3.0
django-cors-headers~=4.9.0
packaging~=25.0
django_ipware~=7.0.1
requests~=2.32.5
fpdf~=1.7.2
django-q2~=1.9.0
blessed~=1.25.0
psycopg[binary]~=3.3.2
whitenoise~=6.11.0
django-storages[dropbox,s3]~=1.14.6
django-htmx~=1.27.0
python-gedcom~=1.0.0

